Adding Firebase (#2)
Adding Firebase to the SvelteKit setup using the adapter. This enables SSR using Cloud Functions as well as Hosting and Emulators.
Changes include configurations of the adapter and of Vite for new versions of Kit. Also initialisation of the function.
I ended up opting out of ESLint for Cloud Functions. This was because the linter didn't agree with the ES standards that the Firebase-Svelte adapter was using for the SSR-function. The linter wasn't providing value in this case, and was only blocking the deploy.